Server/requirements.txt
flask
opencv-python
pillow
numpy
scikit-image
matplotlib

